[GILJOB-66] 퀘스트 수 조회 컨트롤러 테스트 추가
