+++ src/test/kotlin/com/yapp/giljob/domain/quest/api/QuestParticipationControllerTest.kt
@@ -2,10 +2,12 @@ package com.yapp.giljob.domain.quest.api
 
 import com.yapp.giljob.domain.quest.application.QuestParticipationService
 import com.yapp.giljob.domain.quest.dao.QuestParticipationRepository
+import com.yapp.giljob.domain.quest.dto.response.QuestCountResponseDto
 import com.yapp.giljob.domain.user.dao.UserRepository
 import com.yapp.giljob.global.AbstractRestDocs
 import com.yapp.giljob.global.config.security.GiljobTestUser
 import org.junit.jupiter.api.Test
+import org.mockito.BDDMockito
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.boot.test.mock.mockito.MockBean
 import org.springframework.restdocs.headers.HeaderDocumentation
@@ -53,6 +55,39 @@ internal class QuestParticipationControllerTest : AbstractRestDocs() {
                             .description("성공 메세지"),
                         PayloadDocumentation.fieldWithPath("data")
                             .description("null")
+                    )
+                )
+            )
+    }
+
+    @Test
+    fun getAllQuestCountTest() {
+        BDDMockito.given(questParticipationService.getAllQuestCount()).willReturn(QuestCountResponseDto(124L, 27L, 52L))
+
+        val result = mockMvc.perform(RestDocumentationRequestBuilders.get("/api/quests/count")).andDo(
+            MockMvcResultHandlers.print()
+        )
+
+        result
+            .andExpect(MockMvcResultMatchers.status().isOk)
+            .andDo(
+                MockMvcRestDocumentation.document(
+                    "quests/count/get",
+                    HeaderDocumentation.responseHeaders(),
+                    HeaderDocumentation.responseHeaders(),
+                    PayloadDocumentation.responseFields(
+                        PayloadDocumentation.fieldWithPath("status")
+                            .description("200"),
+                        PayloadDocumentation.fieldWithPath("message")
+                            .description("성공 메세지"),
+                        PayloadDocumentation.fieldWithPath("data")
+                            .description("응답 데이터"),
+                        PayloadDocumentation.fieldWithPath("data.totalQuestCount")
+                            .description("전체 퀘스트 수"),
+                        PayloadDocumentation.fieldWithPath("data.onProgressQuestCount")
+                            .description("진행 중 퀘스트 수"),
+                        PayloadDocumentation.fieldWithPath("data.totalParticipantCount")
+                            .description("퀘스트 참여자 수"),
                     )
                 )
             )
